server/requirements.txt
anyio==3.6.1
asgiref==3.5.2
backports.cached-property==1.0.1
boto3==1.21.41
click==8.1.3
python-dotenv==0.20.0
fastapi==0.78.0
graphql-core==3.2.1
greenlet==1.1.2
h11==0.13.0
idna==3.3
matplotlib==3.4.3
motor==3.0.0
numpy==1.20.3
opencv-python==4.5.5.64
pydantic==1.9.1
Pygments==2.12.0
pymongo==4.1.1
python-dateutil==2.8.2
python-multipart==0.0.5
passlib==1.7.4
six==1.16.0
sniffio==1.2.0
SQLAlchemy==1.4.39
starlette==0.19.1
strawberry-graphql==0.114.0
typing_extensions==4.2.0
uvicorn==0.17.6
imutils==0.5.4
Pillow==8.4.0
PyJWT==2.4.0
pytz==2022.1